refactor: move master/minichef addresses to sushi pkg
