+++ apps/web/src/lib/wagmi/hooks/master-chef/use-master-chef.ts
@@ -6,6 +6,11 @@ import { keepPreviousData, useQueryClient } from '@tanstack/react-query'
 import { useCallback, useEffect, useMemo } from 'react'
 import { erc20Abi, masterChefV2Abi, miniChefV2Abi } from 'sushi/abi'
 import { ChainId } from 'sushi/chain'
+import {
+  MASTERCHEF_ADDRESS,
+  MASTERCHEF_V2_ADDRESS,
+  MINICHEF_ADDRESS,
+} from 'sushi/config'
 import { SUSHI, SUSHI_ADDRESS } from 'sushi/currency'
 import { Amount, Token } from 'sushi/currency'
 import { Address, UserRejectedRequestError, encodeFunctionData } from 'viem'
@@ -18,12 +23,7 @@ import {
 } from 'wagmi'
 import { SendTransactionErrorType } from 'wagmi/actions'
 import { SendTransactionData } from 'wagmi/query'
-import {
-  MASTERCHEF_ADDRESS,
-  MASTERCHEF_V2_ADDRESS,
-  MINICHEF_ADDRESS,
-  useMasterChefContract,
-} from './use-master-chef-contract'
+import { useMasterChefContract } from './use-master-chef-contract'
 
 interface UseMasterChefReturn
   extends Pick<ReturnType<typeof useReadContracts>, 'isLoading' | 'isError'> {
