+++ jobs/pool/src/lib/chefs/masterChefV1/index.ts
@@ -2,8 +2,8 @@ import { daysInYear, secondsInDay } from 'date-fns/constants'
 import { ChainId } from 'sushi/chain'
 import { SUSHI } from 'sushi/currency'
 
+import { MASTERCHEF_ADDRESS } from 'sushi/config'
 import { SECONDS_BETWEEN_BLOCKS } from 'sushi/config/subgraph'
-import { MASTERCHEF_ADDRESS } from '../../../config.js'
 import { getPairs, getTokenBalancesOf, getTokens } from '../../common/index.js'
 import type { ChefReturn, Farm } from '../../types.js'
 import { getPoolInfos, getPoolLength, getTotalAllocPoint } from './fetchers.js'
