+++ packages/sushi/src/config/index.ts
@@ -7,6 +7,7 @@ export * from './default-quote.js'
 export * from './furo.js'
 export * from './default-slippage.js'
 export * from './lsds.js'
+export * from './master-chef.js'
 export * from './mini-chef.js'
 export * from './native-ids.js'
 export * from './extractor.js'
