+++ packages/sushi/src/config/master-chef.ts
@@ -1,0 +1,13 @@
+import { ChainId } from '../chain/index.js'
+
+export const MASTERCHEF_ADDRESS = {
+  [ChainId.ETHEREUM]: '0xc2EdaD668740f1aA35E4D8f227fB8E17dcA888Cd',
+  // [ChainId.ROPSTEN]: '0x80C7DD17B01855a6D2347444a0FCC36136a314de',
+  // [ChainId.RINKEBY]: '0x80C7DD17B01855a6D2347444a0FCC36136a314de',
+  // [ChainId.GÖRLI]: '0x80C7DD17B01855a6D2347444a0FCC36136a314de',
+  // [ChainId.KOVAN]: '0x80C7DD17B01855a6D2347444a0FCC36136a314de',
+} as const
+
+export const MASTERCHEF_V2_ADDRESS = {
+  [ChainId.ETHEREUM]: '0xEF0881eC094552b2e128Cf945EF17a6752B4Ec5d',
+} as const
